Created ui_utils.py to refactor app.py
